chore: rename config.py to config_constants.py
